mongo_service/requirements.txt
pymongo==4.6.1
python-dotenv==1.0.1
pydantic==2.5.0

